sre/requirements.txt
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
prometheus-client==0.20.0
requests==2.32.3
PyYAML==6.0.1
slack-sdk==3.31.0
pytest==8.3.2
pytest-cov==5.0.0
responses==0.25.3
mypy==1.11.1
types-requests==2.32.0
types-PyYAML==6.0.12
hypothesis==6.111.1
